Move state and country validation from a trait to a validator class
